Change functions to meet conventions

+++ triviaProject/Project1/Responses.h
@@ -99,7 +99,7 @@ struct PlayerResults
 	string username;
 	unsigned int correctAnswerCount;
 	unsigned int wrongAnswerCount;
-	float averageAnswerCount;
+	double averageAnswerCount;
 };
 
 struct GetGameResultsResponse : public Response
